Moved view services to integration assembly:
- Dumb down the view logic and make the controller responsible for the response
